Add files via upload
In this project I created a fortune teller. It ask's you your mood and then displays an emoji based on your selection. After that it asks you to ask a yes or no question. I set it up with a random number generator to answer, so everything is 50/50... Completely up to the universe. :)